+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/PaymentSheetViewModel.kt
@@ -22,6 +22,7 @@ import com.stripe.android.core.injection.IOContext
 import com.stripe.android.core.injection.Injectable
 import com.stripe.android.core.injection.InjectorKey
 import com.stripe.android.core.injection.injectWithFallback
+import com.stripe.android.features.isEnabled
 import com.stripe.android.googlepaylauncher.GooglePayEnvironment
 import com.stripe.android.googlepaylauncher.GooglePayPaymentMethodLauncher
 import com.stripe.android.googlepaylauncher.GooglePayPaymentMethodLauncherContract
@@ -31,7 +32,6 @@ import com.stripe.android.link.LinkActivityResult
 import com.stripe.android.link.LinkActivityResult.Canceled.Reason
 import com.stripe.android.link.LinkPaymentDetails
 import com.stripe.android.link.LinkPaymentLauncher
-import com.stripe.android.link.LinkPaymentLauncher.Companion.LINK_ENABLED
 import com.stripe.android.link.injection.LinkPaymentLauncherFactory
 import com.stripe.android.link.model.AccountStatus
 import com.stripe.android.model.ConfirmPaymentIntentParams
@@ -435,7 +435,7 @@ internal class PaymentSheetViewModel @Inject internal constructor(
     }
 
     override fun setupLink(stripeIntent: StripeIntent) {
-        if (LINK_ENABLED &&
+        if (PaymentSheetFeatures.link.isEnabled &&
             stripeIntent.paymentMethodTypes.contains(PaymentMethod.Type.Link.code) &&
             stripeIntent.linkFundingSources.intersect(LinkPaymentLauncher.supportedFundingSources)
                 .isNotEmpty()
